requirements.txt
#requirements.txt
accelerate==0.27.0
black==24.2.0
Chroma==0.2.0
ctransformers==0.2.27
faiss-cpu==1.7.4
hf_transfer==0.1.5
httptools==0.6.1
ipykernel==6.29.5
langchain-chroma==0.1.2
langchain-community==0.2.7
langchain-upstage==0.1.7
llama_cpp_python==0.2.40
llava==0.0.1.dev0
loguru==0.7.2
pdfminer.six==20231228
pip==24.1.2
PyMuPDF==1.24.7
python-dotenv==1.0.1
sentence-transformers==2.3.1
streamlit==1.31.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0

